Added the option to grab Title Name or Domain Name with Popup #2
I modified the Manifest.Json and added a options html that works as the popup when icon is clicked as well as added the js for grab title name and grab domain name to be enabled by a button click, added a css page i believe you deleted previously to add to the popup. I tested it on my side with dev mode and it worked properly if you have any question and want to hop on a call my discord is Mixvi.